Update final image size: do it only on darkroom leave
That function only updates the final image size in metadata, for collection purposes. But it is called in many places, assumes we have a darkroom pipeline, and lock its mutex. This hypothesis is a bit wild and the mutex locking will create dead locks. All in all, we only need to update sizes once, when leaving darkroom. Or when we update an history (pasting/styles) in lighttable... but current design anyway works only with an active darkroom pipeline.

All in all, image dimensions metadata are rarely used, we don't really care about them, and anyway they are not important enough to disrupt the whole running sequence.

Another great piece of design from 2020…