Image ID: handle them everywhere as int32_t
imgid is an incremental index set by SQLite, which does not have unsigned integer type, so it's always int at input. We sometimes deal with it as signed, sometimes as unsigned integer. Problem is some parts of the code use imgid = -1 as a flag for unknown/invalid images. If that goes through a part of the code that casts it as uint, then the test imgid > -1 is always TRUE and the result of those checkes is undefined.

Anyway, fix up that mess.