De-implement image ratio and final width/height writing to DB
Those are not image metadata but properties of an export pipeline under user parameters. Problem is there are many ways to manipulate an image history that will affect these properties, and not all of them trigger pipeline recomputes, especially if we copy-paste or apply styles from outside darkroom. Plus the ratio is rounded to one decimal, which may be an issue if/when we compute it over the preview pipeline (720×450 px) or on thumbnails (from lighttable). So, all in all, this is unreliable, badly implemented and half broken.