Collections: handle group **display** in **GUI** code instead of SQL
Once again, the good old view-model-controller paradigm got fucked and managing group members at the collection level led to complications in both GUI and selection code.

This makes "select all" and "invert selection" select all images from the collection, including group members, whether they are displayed or not. It also breaks the range selection.

Since selection is a GUI event that gets saved to DB, and the GUI is known to contain at most all the items of a collection, selecting should not read collection or DB through SQL but directly get a list of imgid from GUI thumbtable.