Remove all traces of custom ordering in collections
Tag collections were hacked with a feature that allows user to reorder images, but that works only IF the current collection is based on tag AND the filtering order is custom AND by drag & dropping from lighttable.

Needless to say, most users don't know it exists, it's undiscoverable in GUI, and anyway this is the worst design for that kind of feature (collections are meant to be SQL extractions from DB).